pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "self-snn"
version = "0.0.1"
description = "Self-agency spiking neural network (self-snn)."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "self-snn contributors"}]
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "numpy",
    "scipy",
    "pyyaml",
    "tqdm",
    "matplotlib",
    "tensorboard",
    "scikit-learn",
    "spikingjelly"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["self_snn*"]

